# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: HerLoop API
  description: |
    HerLoop 她循环 — 女性社群闲置交易平台后端接口。
    所有接口返回 `{code, message, data}` 格式。分页接口使用 `PageResult<T>` 包装。
  version: 1.0.0
  contact:
    name: HerLoop Team
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      tags: [Product]
      summary: 商品列表（公开）
      security: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [全部, 服饰, 图书, 美妆, 运动, 生活]
        - name: tradeMode
          in: query
          schema:
            type: string
            enum: [cash, points, both]
        - name: status
          in: query
          schema:
            type: string
            enum: [on, reserved, sold]
        - name: keyword
          in: query
          schema:
            type: string
          description: 关键词搜索（FULLTEXT on title+description）
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultProductVO"
    post:
      tags: [Product]
      summary: 发布商品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
      responses:
        "200":
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProductVO"
  /products/{id}:
    get:
      tags: [Product]
      summary: 商品详情（公开）
      security: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 商品详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProductVO"
    delete:
      tags: [Product]
      summary: 删除商品（逻辑删除）
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /products/{id}/status:
    put:
      tags: [Product]
      summary: 更新商品状态
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [on, reserved, sold]
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /products/{id}/images:
    post:
      tags: [Product]
      summary: 上传商品图片（最多 5 张）
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
      responses:
        "200":
          description: 上传成功，返回图片 URL 列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: string

  # ── Want ────────────────────────────────────────────────
...
  /trades:
    post:
      tags: [Trade]
      summary: 创建交易
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeCreateRequest"
      responses:
        "200":
          description: 交易创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TradeVO"
  /trades/{id}/complete:
    put:
      tags: [Trade]
      summary: 完成交易
      description: 双方各获得 +50 积分，触发邀请奖励。
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 交易完成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /trades/{id}/cancel:
    put:
      tags: [Trade]
      summary: 取消交易
      description: 退还已扣除积分。
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 交易取消
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # ── Points ──────────────────────────────────────────────
...
  /reports:
    post:
      tags: [Report]
      summary: 提交举报
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreateRequest"
      responses:
        "200":
          description: 举报提交成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Report"
  /reports/pending:
    get:
      tags: [Report]
      summary: 待审核举报列表
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 举报列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/Report"
                          total:
                            type: integer
  /reports/{id}/audit:
    put:
      tags: [Report]
      summary: 审核举报
      parameters:
        - $ref: "#/components/parameters/idPath"
        - name: action
          in: query
          required: true
          schema:
            type: string
          description: 审核操作（confirm / dismiss）
      responses:
        "200":
          description: 审核完成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # ── Admin ───────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
...
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true

    # ── Auth ──
...
    ProductCreateRequest:
      type: object
      required: [title, category, condition, description, tradeMode]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [服饰, 图书, 美妆, 运动, 生活]
        condition:
          type: string
          example: 95新
        description:
          type: string
        cashPrice:
          type: number
        pointsPrice:
          type: integer
        tradeMode:
          type: string
          enum: [cash, points, both]
        icon:
          type: string
        wechat:
          type: string
    ProductVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        condition:
          type: string
        description:
          type: string
        cashPrice:
          type: number
        pointsPrice:
          type: integer
        tradeMode:
          type: string
          enum: [cash, points, both]
        status:
          type: string
          enum: [on, reserved, sold]
        icon:
          type: string
        imageUrls:
          type: string
          description: 逗号分隔的图片 URL
        createdAt:
          type: string
          format: date-time
        seller:
          $ref: "#/components/schemas/SellerBrief"
    SellerBrief:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        verified:
          type: boolean

    # ── Want ──
...
    TradeCreateRequest:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
    TradeVO:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/ProductVO"
        type:
          type: string
          enum: [buy, sell]
        status:
          type: string
        pointsPaid:
          type: integer
          description: 本次交易消耗的积分
        tradeDate:
          type: string
          format: date-time

    # ── Points ──
...
    ReportCreateRequest:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [user, product]
        targetId:
          type: integer
        reason:
          type: string
    Report:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        targetType:
          type: string
        targetId:
          type: integer
        reason:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Admin ──
...
    PageResultProductVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/ProductVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...